pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-stock-playlist-watcher"
version = "0.1.0"
description = "Watch a YouTube playlist and summarize stock-related mentions from new videos."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Project Owner" }
]
dependencies = [
  "google-genai>=1.0.0,<2",
  "youtube-transcript-api>=1.2.4,<2",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
playlist-watcher = "playlist_watcher.main:main"
